Fix handling of repeated letters marked absent
The previous logic incorrectly assumed that if a letter is marked absent, it cannot be present in the word. Instead, we must distinguish whether the letter was also marked correct or present in the same word. If so, then the number of occurrences in the word is
limited to those already reported correct or present.

This required a significant change in the guessing logic, and corresponding changes in the test logic to generate responses from a memorized secret word.